+++ __init__.py
@@ -2,3 +2,21 @@ import sys
 import os
 
 sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
+
+from .lib.core.api import (
+    DirsearchFuzzer,
+    FuzzerConfig,
+    FuzzerResult,
+    Wordlist,
+    WordlistState,
+    WordlistTemplate,
+)
+
+__all__ = [
+    "DirsearchFuzzer",
+    "FuzzerConfig",
+    "FuzzerResult",
+    "Wordlist",
+    "WordlistState",
+    "WordlistTemplate",
+]